Fix RMM C++ sources and tests for MSVC

+++ cpp/tests/mr/statistics_mr_tests.cpp
@@ -208,7 +208,7 @@ TEST(StatisticsTest, NegativeInnerTracking)
   EXPECT_EQ(mr.get_allocations_counter().value, 0);
 
   // The inner_mr will have negative values
-  EXPECT_EQ(inner_mr.get_bytes_counter().value, -100_MiB);
+  EXPECT_EQ(inner_mr.get_bytes_counter().value, -static_cast<int64_t>(100_MiB));
   EXPECT_EQ(inner_mr.get_allocations_counter().value, -10);
 
   // Verify the peak and total
